+++ pylatex/__init__.py
@@ -9,12 +9,14 @@ from .basic import HugeText, NewPage, LineBreak, NewLine, HFill, LargeText, \
     MediumText, SmallText, FootnoteText, TextColor
 from .document import Document
 from .frames import MdFramed, FBox
-from .math import Math, VectorName, Matrix
+from .math import Math, VectorName, Matrix, Alignat
 from .package import Package
 from .section import Section, Subsection, Subsubsection
 from .table import Table, MultiColumn, MultiRow, Tabular, Tabu, LongTable, \
     LongTabu, Tabularx, LongTabularx, ColumnType
-from .tikz import TikZ, Axis, Plot
+from .tikz import TikZ, Axis, Plot, TikZNode, TikZDraw, TikZCoordinate, \
+    TikZPathList, TikZPath, TikZUserPath, TikZOptions, TikZNodeAnchor, \
+    TikZScope
 from .figure import Figure, SubFigure, StandAloneGraphic
 from .lists import Enumerate, Itemize, Description
 from .quantities import Quantity