+++ pylatex/base_classes/latex_object.py
@@ -39,6 +39,7 @@ class LatexObject(metaclass=_CreatePackages):
     """
 
     _latex_name = None
+    _star_latex_name = False    # latex_name + ('*' if True else '')
 
     #: Set this to an iterable to override the list of default repr
     #: attributes.
@@ -55,7 +56,7 @@ class LatexObject(metaclass=_CreatePackages):
 
     @property
     def escape(self):
-        """Determine wheter or not to escape content of this class.
+        """Determine whether or not to escape content of this class.
 
         This defaults to `True` for most classes.
         """
@@ -64,6 +65,11 @@ class LatexObject(metaclass=_CreatePackages):
         if self._default_escape is not None:
             return self._default_escape
         return True
+
+    @escape.setter
+    def escape(self, value):
+        """Escape flag setter - to be used at object level."""
+        self._escape = value
 
     #: Start a new paragraph before this environment.
     begin_paragraph = False
@@ -117,13 +123,14 @@ class LatexObject(metaclass=_CreatePackages):
 
     @property
     def latex_name(self):
-        """The name of the class used in LaTeX.
+        """Return the name of the class used in LaTeX.
 
         It can be `None` when the class doesn't have a name.
         """
+        star = ('*' if self._star_latex_name else '')
         if self._latex_name is not None:
-            return self._latex_name
-        return self.__class__.__name__.lower()
+            return self._latex_name + star
+        return self.__class__.__name__.lower() + star
 
     @latex_name.setter
     def latex_name(self, value):