+++ pylatex/table.py
@@ -377,6 +377,76 @@ class Tabu(Tabular):
 
     packages = [Package('tabu')]
 
+    def __init__(self, table_spec, data=None, pos=None, *,
+                 row_height=None, col_space=None, width=None, booktabs=None,
+                 spread=None, to=None, **kwargs):
+        """
+        Args
+        ----
+        table_spec: str
+            A string that represents how many columns a table should have and
+            if it should contain vertical lines and where.
+        pos: list
+        row_height: float
+            Specifies the heights of the rows in relation to the default
+            row height
+        col_space: str
+            Specifies the spacing between table columns
+        booktabs: bool
+            Enable or disable booktabs style tables. These tables generally
+            look nicer than regular tables. If this is `None` it will use the
+            value of the ``booktabs`` attribte from the `~.active`
+            configuration. This attribute is `False` by default.
+        spread: str
+            Specifies the Tabu table should add a given amount of 'padding' to
+            the width of the table. This should be a latex dimension;
+            for example: "0 pt" or "1in"
+        to: str
+            Specifies the Tabu table should extend to a given width.
+            This should be a latex dimension; for example '4in'
+        width: int
+            The amount of columns that the table has. If this is `None` it is
+            calculated based on the ``table_spec``, but this is only works for
+            simple specs. In cases where this calculation is wrong override the
+            width using this argument.
+
+        References
+        ----------
+        * https://en.wikibooks.org/wiki/LaTeX/Tables#The_tabular_environment
+        """
+
+        super().__init__(table_spec, data, pos,
+                         row_height=row_height, col_space=col_space,
+                         width=width, booktabs=booktabs, **kwargs)
+
+        self._preamble = ""
+        if spread:
+            self._preamble = "spread " + spread
+        elif to:
+            self._preamble = "to " + to
+
+    def dumps(self):
+        """Turn the tabu object into a string in Latex format."""
+
+        _s = super().dumps()
+
+        # Tabu tables support a unusual syntax:
+        # \begin{tabu} spread 0pt {<col format...>}
+        #
+        # Since this syntax isn't common, it doesn't make
+        # sense to support it in the baseclass (e.g., Environment)
+        # rather, here we fix the LaTeX string post-hoc
+        if self._preamble:
+            if _s.startswith(r"\begin{longtabu}"):
+                _s = _s[:16] + self._preamble + _s[16:]
+            elif _s.startswith(r"\begin{tabu}"):
+                _s = _s[:12] + self._preamble + _s[12:]
+            else:
+                raise TableError("Can't apply preamble to Tabu table "
+                                 "(unexpected initial command sequence)")
+
+        return _s
+
 
 class LongTable(Tabular):
     """A class that represents a longtable (multipage table)."""
@@ -384,6 +454,8 @@ class LongTable(Tabular):
     packages = [Package('longtable')]
 
     header = False
+    foot = False
+    lastFoot = False
 
     def end_table_header(self):
         r"""End the table header which will appear on every page."""
@@ -394,7 +466,29 @@ class LongTable(Tabular):
 
         self.header = True
 
-        self.append(NoEscape(r'\endhead'))
+        self.append(Command(r'endhead'))
+
+    def end_table_footer(self):
+        r"""End the table foot which will appear on every page."""
+
+        if self.foot:
+            msg = "Table already has a foot"
+            raise TableError(msg)
+
+        self.foot = True
+
+        self.append(Command('endfoot'))
+
+    def end_table_last_footer(self):
+        r"""End the table foot which will appear on the last page."""
+
+        if self.lastFoot:
+            msg = "Table already has a last foot"
+            raise TableError(msg)
+
+        self.lastFoot = True
+
+        self.append(Command('endlastfoot'))
 
 
 class LongTabu(LongTable, Tabu):