+++ testall.sh
@@ -66,20 +66,27 @@ else
 fi
 
 echo -e '\e[32mTesting tests directory\e[0m'
-if ! $python $(which nosetests) tests/*; then
+if ! $python $(which nosetests) --with-coverage tests/*; then
     exit 1
 fi
+mv .coverage{,.tests}
 
 if [ "$python_version" = '2' ]; then
     cd ..
 fi
 
+
+count=0
 for f in $main_folder/examples/*.py; do
     echo -e '\e[32mTesting '$f'\e[0m'
-    if ! $python $f; then
+    if ! $python $(which coverage) run $f; then
         exit 1
     fi
+    ((count ++))
+    mv .coverage .coverage.example$count
 done
+
+coverage combine
 
 if [ "$clean" = 'TRUE' ]; then
     rm *.pdf *.log *.aux *.tex *.fls *.fdb_latexmk > /dev/null
@@ -89,8 +96,10 @@ fi
 if [[ "$nodoc" != 'TRUE' && "$python_version" == "3" && "$python_version_long" != 3.3.* ]]; then
     echo -e '\e[32mChecking for errors in docs and docstrings\e[0m'
     cd docs
+    set -e
     ./create_doc_files.sh -p $python
     make clean
+    set +e
     if ! $python $(which sphinx-build) -b html -d build/doctrees/ source build/html -nW; then
         exit 1
     fi